multimodal-rag-agents/pyproject.toml
[project]
name = "multimodal-rag-agents"
version = "0.1.0"
description = "Multi-Agent Multimodal RAG System with OpenAI and Instructor"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "openai>=1.86.0",
    "instructor>=1.7.2",
    "pydantic>=2.8.0",
    "voyageai>=0.3.2",
    "astrapy>=2.0.1",
    "pillow>=11.2.1",
    "pymupdf>=1.26.1",
    "pymupdf4llm>=0.0.24",
    "python-dotenv>=1.1.0",
    "tqdm>=4.67.1",
    "asyncio-pool>=0.7.0",
    "tenacity>=9.1.2",
]

[project.optional-dependencies]
test = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "pytest-mock>=3.14.0",
    "coverage>=7.6.0",
]
dev = [
    "multimodal-rag-agents[test]",
    "black>=24.0.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/rag_agents"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "--verbose",
]

[tool.coverage.run]
source = ["src"]
branch = true

[tool.coverage.report]
show_missing = true
skip_covered = false
fail_under = 80